src/email_watcher/requirements.txt
python-dotenv
requests
pdfplumber
pillow
pyzbar
beautifulsoup4
lxml
xmltodict
imapclient
O365
tabulate
